Poting to Windows 64 bits.  We're down from 606 to 38 warnings. There are still quite a few warnings in logger.mx which does weird things converting oid to bat and v.v. There is also still a problem with the MT_Id type which is defined as size_t and converted to int. Some built-in commands now take wrd argument but there are MIL procs in place with int args for backward compatibility. I'm adding "64bit" comments where a closer look is warranted.

+++ MonetDB/src/tools/prog.c
@@ -168,7 +168,7 @@ main(int argc, char **av)
 			break;
 		}
 		while ((line = fgets(buf+curlen, 1024, fp)) != NULL) {
-			int n = strlen(line);
+			int n = (int) strlen(line);
             		curlen += n;
             		if (curlen+1024 > maxlen) {
                			maxlen += 8*BUFSIZ;
